FIX KB-26: Moved everything in `AuthContents` to `AuthorisationPage`
