+++ frontend/toge-trades/src/Pages/Authorisation/AuthorisationPage.css
@@ -29,6 +29,89 @@
     background-color: #C9DCF4;
     z-index: -100;
 }
+.content-container{
+    position: absolute;
+    height: 100%;
+    width: 45%;
+    top: 0;
+    right: 0;
+    transition: 0.8s ease-in-out; 
+}
+.signup-screen{
+    transform: translateX(-128%);
+}
+.login-screen{
+    transform: translateX(0%);
+}
+.contents-auth{
+    padding: 1.3em;
+}
+.auth-heading{
+    width: 100%;
+    text-align: left;
+}
+.auth-heading .title{
+    font-size: 2em;
+}
+.subtitle{
+    font-size: 1em;
+    color: #616980;
+}
+.subtitle .auth-link{
+    color: #3B5BBA;
+    font-weight: 700;
+}
+.auth-forms{
+  width: 100%; 
+
+}
+.login-form{
+    opacity: 0;
+    pointer-events: none;
+}
+.signup-form{
+    opacity: 1;
+    pointer-events: all;
+}
+.input-container{
+    padding-top: 0.5em;
+    padding-bottom: 1.5em;
+    transition: opacity 0.02s 0.4s;
+}
+.input-container input{
+    width: 100%;
+    height: 100%;
+    background-color: transparent;
+    border: none;
+    border-bottom: 1px solid #7F869B;
+    padding: 0.5em;
+    color: #7F869B;
+    font-size: 1.2em;
+}
+input:focus {
+    outline: none; /* Removes the default focus border */
+}
+
+.auth-forms button{
+    width: 100%;
+}
+
+@media screen and (max-width: 750px) {
+    .content-container{
+        position:unset;
+        width:100%;
+        height:auto;
+    }
+    .signup-screen{
+        /* No transtions for mobile */
+        transform: translateX(0%);
+    }
+    .input-container{
+        padding-top: 0.5em;
+        padding-bottom: 1.5em;
+        transition: opacity 0.4s ease-in-out;
+    }
+}
 
 @media screen and (max-width: 750px) {
     .auth-box{
